# src/RexSoftwareTest/ApiBundle/Resources/public/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is the \"Movies\" api.\n",
    "version": "1.0.0",
    "title": "Movies API"
  },
  "host": "symfony.dev",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/oauth/v2/token": {
      "post": {
        "summary": "Authenticate via username and password.",
        "description": "",
        "tags": [
          "Auth"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "credentials",
            "description": "{ \"grant_type\": \"password\",\r\n \"client_id\": \"1_3bcbxd9e24g0gk4swg0kwgcwg4o8k8g4g888kwc44gcc0gwwk4\",\r\n \"client_secret\": \"4ok2x70rlfokc8g0wws8c8kwcokw80k44sg48goc0ok4w0so0k\",\r\n \"username\": \"admin\",\r\n \"password\": \"today123\"}\r\n",
            "required": true,
            "schema": {
              "$ref": "#/definitions/OAuthCredentials"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/": {
      "get": {
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "description": "",
        "tags": [
          "Swagger"
        ],
        "summary": "Get the swagger doc for the API.",
        "operationId": "get_api",
        "security": [
          {
            "apiKey": []
          }
        ]
      }
    },
    "/api/movies": {
      "get": {
        "responses": {
          "200": {
            "description": "All movies that meet the conditions are returned.",
            "schema": {
              "$ref": "#/definitions/RexSoftwareTest.ApiBundle.Entity.Movie[]"
            }
          }
        },
        "description": "",
        "tags": [
          "Movie"
        ],
        "summary": "Get all movies that optionally meet certain conditions.",
        "operationId": "get_api-movies",
        "security": [
          {
            "apiKey": []
          }
        ]
      },
      "post": {
        "responses": {
          "200": {
            "description": "A new movie was created and returned.",
            "schema": {
              "$ref": "#/definitions/RexSoftwareTest.ApiBundle.Entity.Movie"
            }
          },
          "400": {
            "description": "A bad request was made, any errors will be returned.",
            "schema": {
              "$ref": "#/definitions/RexSoftwareTest.ApiBundle.Model.HttpValidationErrorModel"
            }
          }
        },
        "description": "",
        "tags": [
          "Movie"
        ],
        "summary": "Create a new movie, and get the (now fully initialized) movie, with it's new id.",
        "operationId": "post_api-movies",
        "parameters": [
          {
            "in": "body",
            "description": "",
            "name": "rexsoftwaretest_apibundle_movie",
            "required": false,
            "schema": {
              "$ref": "#/definitions/RexSoftwareTest.ApiBundle.Form.MovieType"
            }
          }
        ],
        "security": [
          {
            "apiKey": []
          }
        ]
      }
    },
    "/api/movies/{movie}": {
      "delete": {
        "responses": {
          "200": {
            "description": "The movie was deleted, and it's last value returned.",
            "schema": {
              "$ref": "#/definitions/RexSoftwareTest.ApiBundle.Entity.Movie"
            }
          },
          "404": {
            "description": "The movie was not found, an error message will be returned.",
            "schema": {
              "$ref": "#/definitions/RexSoftwareTest.ApiBundle.Model.HttpErrorResponseModel"
            }
          }
        },
        "description": "",
        "tags": [
          "Movie"
        ],
        "summary": "Delete a given movie, and return it's last value.",
        "operationId": "delete_api-movies-movie",
        "parameters": [
          {
            "in": "path",
            "name": "movie",
            "required": true,
            "type": "string"
          }
        ],
        "security": [
          {
            "apiKey": []
          }
        ]
      },
      "get": {
        "responses": {
          "200": {
            "description": "The movie with the id provided is returned.",
            "schema": {
              "$ref": "#/definitions/RexSoftwareTest.ApiBundle.Entity.Movie"
            }
          },
          "404": {
            "description": "The movie was not found, an error message will be returned.",
            "schema": {
              "$ref": "#/definitions/RexSoftwareTest.ApiBundle.Model.HttpErrorResponseModel"
            }
          }
        },
        "description": "",
        "tags": [
          "Movie"
        ],
        "summary": "Get a movie by id.",
        "operationId": "get_api-movies-movie",
        "parameters": [
          {
            "in": "path",
            "name": "movie",
            "required": true,
            "type": "string"
          }
        ],
        "security": [
          {
            "apiKey": []
          }
        ]
      },
      "patch": {
        "responses": {
          "200": {
            "description": "The movie with the given id is returned after being updated with the changes.",
            "schema": {
              "$ref": "#/definitions/RexSoftwareTest.ApiBundle.Entity.Movie"
            }
          },
          "400": {
            "description": "A bad request was made, any errors will be returned.",
            "schema": {
              "$ref": "#/definitions/RexSoftwareTest.ApiBundle.Model.HttpValidationErrorModel"
            }
          },
          "404": {
            "description": "The movie was not found, an error message will be returned.",
            "schema": {
              "$ref": "#/definitions/RexSoftwareTest.ApiBundle.Model.HttpErrorResponseModel"
            }
          }
        },
        "description": "",
        "tags": [
          "Movie"
        ],
        "summary": "Update an existing movie, only updating the provided values.",
        "operationId": "patch_api-movies-movie",
        "parameters": [
          {
            "in": "path",
            "name": "movie",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "description": "",
            "name": "rexsoftwaretest_apibundle_movie",
            "required": false,
            "schema": {
              "$ref": "#/definitions/RexSoftwareTest.ApiBundle.Form.MovieType"
            }
          }
        ],
        "security": [
          {
            "apiKey": []
          }
        ]
      },
      "put": {
        "responses": {
          "200": {
            "description": "The movie with the given id is returned after being updated with the changes.",
            "schema": {
              "$ref": "#/definitions/RexSoftwareTest.ApiBundle.Entity.Movie"
            }
          },
          "400": {
            "description": "A bad request was made, any errors will be returned.",
            "schema": {
              "$ref": "#/definitions/RexSoftwareTest.ApiBundle.Model.HttpValidationErrorModel"
            }
          },
          "404": {
            "description": "The movie was not found, an error message will be returned.",
            "schema": {
              "$ref": "#/definitions/RexSoftwareTest.ApiBundle.Model.HttpErrorResponseModel"
            }
          }
        },
        "description": "",
        "tags": [
          "Movie"
        ],
        "summary": "Update an existing movie, overwriting all values.",
        "operationId": "put_api-movies-movie",
        "parameters": [
          {
            "in": "path",
            "name": "movie",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "description": "",
            "name": "rexsoftwaretest_apibundle_movie",
            "required": false,
            "schema": {
              "$ref": "#/definitions/RexSoftwareTest.ApiBundle.Form.MovieType"
            }
          }
        ],
        "security": [
          {
            "apiKey": []
          }
        ]
      }
    }
  },
  "definitions": {
    "OAuthCredentials": {
      "type": "object",
      "required": [
        "grant_type",
        "client_id",
        "client_secret",
        "username",
        "password"
      ],
      "properties": {
        "grant_type": {
          "type": "string"
        },
        "client_id": {
          "type": "string"
        },
        "client_secret": {
          "type": "string"
        },
        "username": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "RexSoftwareTest.ApiBundle.Entity.Movie": {
      "properties": {
        "id": {
          "type": "integer",
          "description": "integer",
          "format": "int32"
        },
        "name": {
          "type": "string",
          "description": "The title of the movie, non nullable."
        },
        "description": {
          "type": "string",
          "description": "The movie's description, this field is nullable."
        },
        "image": {
          "type": "string",
          "description": "A unique identifier for the movie's image (poster, etc), this field is nullable."
        },
        "rating": {
          "type": "number",
          "description": "The movie's current rating / 10, updated with each rating to avoid excessive re-calculation, it starts at 6...\nor so the original plan was, for simplicity it just gets set, right now.",
          "format": "float"
        },
        "role_ids": {
          "type": "array",
          "description": "The ids of the roles that are in the movie.",
          "items": {
            "type": "integer"
          }
        }
      },
      "type": "object"
    },
    "RexSoftwareTest.ApiBundle.Entity.Movie[]": {
      "type": "array",
      "items": {
        "$ref": "#/definitions/RexSoftwareTest.ApiBundle.Entity.Movie"
      }
    },
    "RexSoftwareTest.ApiBundle.Form.MovieType": {
      "description": "object (MovieType)",
      "properties": {
        "name": {
          "type": "string",
          "description": "string"
        },
        "description": {
          "type": "string",
          "description": "string"
        },
        "image": {
          "type": "string",
          "description": "string"
        },
        "rating": {
          "type": "number",
          "description": "float",
          "format": "float"
        },
        "role_ids": {
          "type": "array",
          "description": "array of integers",
          "items": {
            "type": "integer"
          }
        }
      },
      "required": [
        "name"
      ],
      "type": "object"
    },
    "RexSoftwareTest.ApiBundle.Model.HttpErrorDetailsModel": {
      "description": "The error details.",
      "properties": {
        "code": {
          "type": "integer",
          "description": "The http code.",
          "format": "int32"
        },
        "message": {
          "type": "string",
          "description": "An error message."
        }
      },
      "type": "object"
    },
    "RexSoftwareTest.ApiBundle.Model.FormValidationErrorModel": {
      "description": "The form and errors that caused this response.",
      "properties": {
        "name": {
          "type": "string",
          "description": "string"
        },
        "errors": {
          "type": "array",
          "description": "array of strings",
          "items": {
            "type": "string"
          }
        },
        "children": {
          "type": "array",
          "description": "Any child forms that also contain errors.",
          "items": {
            "$ref": "#/definitions/RexSoftwareTest.ApiBundle.Model.FormValidationErrorModel"
          }
        }
      },
      "type": "object"
    },
    "RexSoftwareTest.ApiBundle.Model.HttpValidationErrorModel": {
      "description": "",
      "properties": {
        "error": {
          "$ref": "#/definitions/RexSoftwareTest.ApiBundle.Model.HttpErrorDetailsModel"
        },
        "form": {
          "$ref": "#/definitions/RexSoftwareTest.ApiBundle.Model.FormValidationErrorModel"
        }
      },
      "type": "object"
    },
    "RexSoftwareTest.ApiBundle.Model.HttpErrorResponseModel": {
      "description": "",
      "properties": {
        "error": {
          "$ref": "#/definitions/RexSoftwareTest.ApiBundle.Model.HttpErrorDetailsModel"
        }
      },
      "type": "object"
    }
  },
  "securityDefinitions": {
    "apiKey": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization"
    }
  },
  "tags": [
    {
      "name": ""
    }
  ]
}